+++ lib/Extension/LanguageServer/LanguageServerExtension.php
@@ -9,6 +9,7 @@ use Phpactor\Container\ContainerBuilder;
 use Phpactor\Container\Extension;
 use Phpactor\Extension\LanguageServer\Dispatcher\PhpactorDispatcherFactory;
 use Phpactor\Extension\LanguageServer\EventDispatcher\LazyAggregateProvider;
+use Phpactor\Extension\LanguageServer\Formatter\NoFormatterConfiguredFormatter;
 use Phpactor\Extension\LanguageServer\Handler\DebugHandler;
 use Phpactor\Extension\LanguageServer\Listener\InvalidConfigListener;
 use Phpactor\Extension\LanguageServer\Listener\SelfDestructListener;
@@ -26,6 +27,7 @@ use Phpactor\LanguageServer\Core\Diagnostics\DiagnosticsEngine;
 use Phpactor\LanguageServer\Diagnostics\CodeActionDiagnosticsProvider;
 use Phpactor\LanguageServer\Handler\System\StatsHandler;
 use Phpactor\LanguageServer\Handler\TextDocument\CodeActionHandler;
+use Phpactor\LanguageServer\Handler\TextDocument\FormattingHandler;
 use Phpactor\LanguageServer\Handler\TextDocument\TextDocumentHandler;
 use Phpactor\LanguageServer\Core\Handler\HandlerMethodRunner;
 use Phpactor\LanguageServer\Adapter\DTL\DTLArgumentResolver;
@@ -78,6 +80,7 @@ class LanguageServerExtension implements Extension
     public const TAG_STATUS_PROVIDER = 'language_server.status_provvder';
     public const TAG_DIAGNOSTICS_PROVIDER = 'language_server.diagnostics_provider';
     public const TAG_MIDDLEWARE = 'language_server.middleware';
+    public const TAG_FORMATTER = 'language_server.formatter';
     public const PARAM_SESSION_PARAMETERS = 'language_server.session_parameters';
     public const PARAM_CLIENT_CAPABILITIES = 'language_server.client_capabilities';
     public const PARAM_ENABLE_WORKPACE = 'language_server.enable_workspace';
@@ -414,6 +417,29 @@ class LanguageServerExtension implements Extension
                 $container->get(self::SERVICE_SESSION_WORKSPACE)
             );
         }, [ self::TAG_METHOD_HANDLER => []]);
+
+        $container->register(FormattingHandler::class, function (Container $container) {
+            $formatter = null;
+            foreach ($container->getServiceIdsForTag(self::TAG_FORMATTER) as $seviceId => $_) {
+                $formatter = $container->get($seviceId);
+                if (null === $formatter) {
+                    continue;
+                }
+                break;
+            }
+
+            if ($formatter === null) {
+                $formatter = new NoFormatterConfiguredFormatter($container->get(ClientApi::class));
+            }
+
+            return new FormattingHandler(
+                $container->get(self::SERVICE_SESSION_WORKSPACE),
+                $formatter
+            );
+        }, [
+            LanguageServerExtension::TAG_METHOD_HANDLER => [
+            ],
+        ]);
     }
 
     private function registerServices(ContainerBuilder $container): void